Correct calculation of pivots
For some reason I had used a spacing of `np-1` between selected
pivots, which means that the local data would have to have at least
`(np-1)^2` points to avoid a segfault. This calculation didn't make
any sense anyway, so it's been adjusted to take the actual size of the
local data into account.