Introduce a new adaptive coordinate
This is a very different form of the adaptive coordinate introduced
last year.  This is a pure flux form equation, with no separate
"vertical grid diffusivity" step.  There are three major terms (of
which one is currently unimplemented):

- a density adaptivity term, where we adjust interface positions based
  on differences in locally-referenced potential density
- a pressure smoothing term, which notably *does not* necessarily flatten
  interfaces according to the geopotential
- a weak restoring (source) term; because the equation conserves the
  integrated layer volumes, we are limited in how we can globally
  adjust the grid. this term allows us to break that conservation if
  required, for example to enforce resolution in the mixed layer

For communication, we're required to expand the halo for T/S prior to
ALE due to the calculation of lateral gradients in the adaptive
regridding routine.  This is currently hardcoded, but should be
determined at runtime to avoid excessive communication in other
coordinates that don't require it.

Alongside the implementation of the coordinate itself, there are some
other fixes along the way to supporting code:

- the accelererated regridding routine are now timestep-aware
- we pass the current timestep through to regridding routines so that
  they may use physical timescales (we use this in a physical
  diffusivity term)
- we can ask the model to save the initial conditions used, even after
  a restart -- this is useful if we're altering the grid from the
  restart by accelerated regridding
- despite being used for the diagnostic code themselves (thus causing
  a cyclic dependency), regridding routines may pass out diagnostic
  information to step_MOM() to be written out