Add "barotropically limited" smoothing to adaptive coord
We had been using fluxes limited by local layer thickness in the
adaptive coordinate.  This technique can end up in a regime where
vanished layers persist and are unable to detangle.  We add a second
contribution to the smoothing flux with a different limiter:

If the diffusivity within a column is constant, it's impossible for
layers to run through each other from the smoothing term.  In this case,
the limiter only needs to consider the local topography and the free
surface.  This gives a much stronger smoothing term that is free to
break regions that are otherwise stuck with the layer-based limiter.