Convert the barotropically-limited smoothing to a biharmonic term
We've observed that the minimum contribution to the smoothing of
interface heights, which is barotropically- rather than locally
limited, can be too strong. Instead, we'll implement it with a
biharmonic operator, which should be more scale-selective.

The implementation here might be a bit inefficient, but it's simple to
understand. The question may arise as to whether the limiters need to
be re-applied, but I think it's correct as it is.

As part of this, the `min_smooth` parameter has now become
`biharmonic_smoothing`, and the density/pressure adaptivity terms now
flip between 0 and 1 (mutually exclusive).