Account for cross-velocity term in barotropic velocity remainder
bt_rem_u and bt_rem_v should contain the fraction of remaining
barotropic velocity, with the rest being lost to bottom drag. Because
of this, they should account for the loss due to the cross-velocity
term too.

Because the cross-velocity term is interpolated from a single field,
I've made the simplification that we can simply reduce bt_rem_u by
lin_drag_uv_u (i.e. the drag interpolated on the u points). Perhaps
this isn't entirely correct, because we actually use lin_drag_uv_v,
with the u-velocity interpolated onto the v-points.