Only calculate cross-velocity bt components on unmasked points
Unexpectedly, G%areaCu and friends can be 0 outside of the grid,
leading to division by zero. We avoid this by only calculating the
cross-velocity bt components on unmasked points, where hopefully the
areas of the adjacent cells are well-defined.

Additionally, just to be sure we also initialise these cross-velocity
components to zero everywhere.