Apply d[uv]btdt independent of real component
The calculation of the Rayleigh components applying to the barotropic
velocity derivatives were in the same block as the Rayleigh component
for wave drag directly. It's very often the case that these don't
coincide, so in effect the spring force was never being applied.

This makes the spring force do something: crash the model. But it's a
step forward. The next task is to figure out:
- whether the derivative calculation is correct;
- how to apply the spring force in the bt acceleration terms