Only compute interpolated bt velocities when we can
We managed to hit a very niche edge case in the interpolation of the
barotropic velocities: where a u-point is unmasked, but all the
surrounding v-points from which we interpolate aren't. Because we're
using area weighting, this leads to a division by zero. For the moment
we just implement the simple solution of ensuring we don't divide by zero.