Add PDE-based wave field tracer
Currently hard-coded to work with the two horizontal velocity
components. May need tweaking with how the forcing is included: I have
full forcing by the impulse response of the filter, but it may only be
a delta function response at the centre of the window. The
mathematics and description seem to contradict each other slightly on
this point.